feat: 서비스 모듈화 완료 - 비즈니스 로직 분리 및 관심사 분리
- 할인 서비스 모듈화: discountService.js 생성
- 포인트 서비스 모듈화: loyaltyService.js 생성
- 장바구니 상태 관리 모듈화: cartState.js 생성
- main.basic.js에서 중복된 함수들 제거 및 모듈 import 적용
